+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added compatibility all the way down to android API level 26 for Java.
+- Added support for enum and collection of enum return types for Java.
 
 ### Changed